avoid string allocation for <s

+++ src/embeds/osu/rank.rs
@@ -89,14 +89,14 @@ impl RankEmbed {
                 let title = format!(
                     "How many pp is {name} missing to reach rank #{rank}?",
                     name = user.username,
-                    rank = with_comma_uint(*rank as u64),
+                    rank = with_comma_uint(*rank),
                 );
 
                 let description = if user_pp > *required_pp {
                     format!(
                         "Rank #{rank} currently requires **{required_pp}pp**, \
                         so {name} is already above that with **{pp}pp**.",
-                        rank = with_comma_uint(*rank as u64),
+                        rank = with_comma_uint(*rank),
                         required_pp = with_comma_float(*required_pp),
                         name = user.username,
                         pp = with_comma_float(user_pp)
@@ -108,7 +108,7 @@ impl RankEmbed {
                         "Rank #{rank} currently requires **{required_pp}pp**, \
                         so {name} is missing **{missing}** raw pp, \
                         achievable with a single score worth **{pp}pp**.",
-                        rank = with_comma_uint(*rank as u64),
+                        rank = with_comma_uint(*rank),
                         required_pp = with_comma_float(*required_pp),
                         name = user.username,
                         missing = with_comma_float(required_pp - user_pp),
@@ -119,7 +119,7 @@ impl RankEmbed {
                         "Rank #{rank} currently requires **{required_pp}pp**, \
                         so {name} is missing **{required_pp}** raw pp, \
                         achievable with a single score worth **{required_pp}pp**.",
-                        rank = with_comma_uint(*rank as u64),
+                        rank = with_comma_uint(*rank),
                         required_pp = with_comma_float(*required_pp),
                         name = user.username,
                     )
